% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{asfr_subset}
\alias{asfr_subset}
\title{Human Fertility Database Data on
Age-Specific Fertility Rates}
\format{
A tibble with 33572 rows and the following
columns: \code{Code}, \code{Year}, \code{Age},\code{ASFR}.
}
\source{
Downloaded from the "Age-specific fertility rate"
row of the "By Statistic" table on the
\href{https://www.humanfertility.org/Data/ZippedDataFiles}{ZippedDataFiles}
page of the Human Fertility Database.
Accessed 12 April 2024.
}
\usage{
asfr_subset
}
\description{
Fertility rates by single year of age,
by county, by calendar year.
from the Human Fertility Database.
}
\details{
\code{asfr_subset} is for testing, rather than data analysis.
It only contains data for the period 1980-2000,
and is not kept up-to-date. Data for analysis
should be obtained from the HFD itself.
}
\keyword{datasets}
