% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssvd-package.R
\docType{package}
\name{bssvd-package}
\alias{bssvd}
\alias{bssvd-package}
\title{bssvd: Make Scaled SVD Objects}
\description{
Build objects holding data on age profiles for demographic processes such as fertility, mortality, and labour force participation. The objects are constructed by performing singular value decompositions on data from international demographic databasaes. Package 'bssvd' is designed to support package 'bage'. It is developer-focused, and would not normally be used directly for data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayesiandemography.github.io/bssvd/}
}

}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

}
\keyword{internal}
