% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{coef_csa}
\alias{coef_csa}
\title{Obtain Coefficients from Scaled SVD of
Census School Attendance Data}
\usage{
coef_csa(data, n_comp = 5, eps = 1e-05)
}
\arguments{
\item{data}{Data frame with attendance data.
(Data from multiple files combined, but otherwise
unprocessed).}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Obtain coefficients from a
scaled SVD of school attendance data.
The attendance data is derived from
population censuses by the United Nations Statistics Division.
The data is available in the
"Population 5 to 24 years of age by school
attendance, sex and urban/rural residence" table in
from the
\href{https://unstats.un.org/unsd/demographic-social/products/dyb/index.cshtml#censusdatasets}{Population Censuses' Datasets}.
The coefficients are a
scaled version of the \eqn{U} matrix
from the SVD.
}
\section{Truncation and transformation}{


Reported proportions outside \verb{(eps, 1-eps)}
are shifted to \code{eps} or \code{1-eps}, and
all values are logit-transformed,
before the SVD is applied.
}

\examples{
coef_csa(un_csa_subset)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_csa]{data_ssvd_csa()}} Put school attendance data
into the format required by
the \code{ssvd()} function in \pkg{bage}
\item \code{\link[=tidy_csa]{tidy_csa()}} Format school attendance data
into a tidy data frame.
}
}
