% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himd.R
\name{coef_himd}
\alias{coef_himd}
\title{Obtain Coefficients from Scaled SVD of HIMD Data}
\usage{
coef_himd(
  zipfile,
  time_interval = 1,
  measure_type = c("rate", "prob"),
  n_comp = 5
)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Internal Migration Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{time_interval}{Length of interval over
which values are calculated. Choices are
\code{1} or \code{5}. Default is \code{1}.}

\item{measure_type}{\code{"prob"} or \code{"rate"}.
Whether values are to be
treated as probabilities or
as rates. See \code{\link[=data_ssvd_himd]{data_ssvd_himd()}} for details.
Default is \code{"rate"}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Obtain time series of coefficients from a
scaled SVD of data from the
\href{https://osf.io/vmrfk/}{Human Internal Migration Database}.
The coefficients are a
scaled version of the \eqn{U} matrix
from the SVD.
}
\examples{
zipfile <- system.file("extdata", "hmd_statistics_subset.zip",
                       package = "bssvd")
coef_hmd(zipfile)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Put data
from the Human Internal Migration Database
into the format required by
the \code{ssvd()} function in \code{bage}
\item \code{\link[=tidy_hmd]{tidy_hmd()}} Format data from the
Human Internal Migration Database into
a tidy data frame
}
}
