% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmd.R
\name{coef_wmd}
\alias{coef_wmd}
\title{Obtain Coefficients from Scaled SVD of
World Marriage Data}
\usage{
coef_wmd(file, n_comp = 5, status = c("current", "ever"), eps = 1e-05)
}
\arguments{
\item{file}{Path to excel spreadsheet
holding World Marriage Data}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}

\item{status}{\code{"current"} (the default)
or \code{"ever"}. Whether to report results for
the marriage staus "currently married"
or "ever-married". Note that both statuses
include consensual and de facto unions,
in addition to legal marriages.}

\item{eps}{Parameter for truncating
probabililities or rates.
Default is \code{0.00001}.}
}
\value{
A tibble
}
\description{
Obtain coefficients from a
scaled SVD of marriage data from the
UN Population Division's
\href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Marriage Data}.
The coefficients are a
scaled version of the \eqn{U} matrix
from the SVD.
}
\section{Truncation and transformation}{


Reported proportions outside \verb{(eps, 1-eps)}
are shifted to \code{eps} or \code{1-eps}, and
all values are logit-transformed,
before the SVD is applied.
}

\examples{
file <- system.file(
  "extdata",
  "undesa_pd_2019_wmd_marital_status_subset.xlsx",
  package = "bssvd"
)
coef_wmd(file)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_wmd]{data_ssvd_wmd()}} Put World Marriage Data
into the format required by
the \code{ssvd()} function in \pkg{bage}
\item \code{\link[=tidy_wmd]{tidy_wmd()}} Format World Marriage Data
into a tidy data frame.
}
}
