% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{country_code}
\alias{country_code}
\title{Country Codes}
\format{
A tibble with 190 rows and with
columns \code{country} and \code{code}.
}
\source{
United Nations Statistics Division. (2024).
Standard country or area codes for statistical
use (M49). United Nations, Department of Economic
and Social Affairs. Retrieved from
https://unstats.un.org/unsd/methodology/m49/
on 13 October 2025.
}
\usage{
country_code
}
\description{
Three-digit country codes, and
country names,
from UN_M49 standard.
}
\keyword{datasets}
