% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{data_ssvd_csa}
\alias{data_ssvd_csa}
\title{Prepare World Marriage Data}
\usage{
data_ssvd_csa(data, n_comp = 5, eps = 1e-05)
}
\arguments{
\item{data}{Data frame with attendance data.
(Data from multiple files combined, but otherwise
unprocessed).}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Process marriage data downloaded from
the UN Statistics Division's
\href{https://unstats.un.org/unsd/demographic-social/products/dyb/index.cshtml#censusdatasets}{Population Censuses' Datasets},
collection so it is ready to create a scaled SVD.
}
\section{Usage}{

\strong{Step 1: Download data}

Go to to page
\href{https://unstats.un.org/unsd/demographic-social/products/dyb/index.cshtml#censusdatasets}{Population Censuses' Datasets},
Go to the table "Population 15 years of age and over by educational attainment, age and sex".
Go to the download tab, and download. Only 100,000 rows can be downloaded at one time
so this needs to be done in two steps.

\strong{Step 2: Combine datasets}

Read in the downloadeded files, and combine into a single
data frame.

\strong{Step 3: Call function \code{data_ssvd_csa()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{csa_data <- data_ssvd_ssvd(data)
}\if{html}{\out{</div>}}
}

\section{coef_csa Truncation and transformation}{
NA
}

\examples{
data_ssvd_csa(un_csa_subset)
}
\seealso{
\itemize{
\item \code{\link[=coef_csa]{coef_csa()}} Obtain time series
of SVD coefficients
from World Marriage Data
\item \code{\link[=tidy_csa]{tidy_csa()}} Format World Marriage Data
into a tidy data frame.
}
}
