% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd.R
\name{data_ssvd_hfd}
\alias{data_ssvd_hfd}
\title{Prepare Data from the Human Fertility Database}
\usage{
data_ssvd_hfd(data, age_min_max = 15, age_max_min = 50, n_comp = 5)
}
\arguments{
\item{data}{A data frame containing HFD data.}

\item{age_min_max, age_max_min}{Every age classification
must at least span the range \verb{[age_min_max, age_max_min)}.
Defaults are \code{15} and \code{50}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Process age-specific fertility data from the
the \href{https://www.humanfertility.org/Home/Index}{Human Fertility Database}
so that it is ready to create a scaled SVD.
}
\section{Usage}{

\strong{Step 1: Download data}

Register or log in at the Human Fertility Database,
and go to page
\href{https://www.humanfertility.org/Data/ZippedDataFiles}{ZippedDataFiles}.
Go to the "By statistic" table, and
download the file from the "Age-specific fertility rate" row.

\strong{Step 2: Extract \code{asfrRR.txt}}

Extract the file \code{asfrRR.txt} from the downloaded data,
and read in the contents,

\if{html}{\out{<div class="sourceCode">}}\preformatted{asfr <- readr::read_table("asfrRR.txt", skip = 2)
}\if{html}{\out{</div>}}

\strong{Step 3: Call function \code{data_ssvd_hfd()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{hfd_data <- data_ssvd_hfd(asfr)
}\if{html}{\out{</div>}}
}

\section{Lowest and highest ages}{


The original HFD data contains age groups \code{"12-"} (ie 12 and younger)
and \code{"55+"} (ie 55 and older). We treat these as closed intervals
\code{"12"} and \code{"55"}.

If an age classification in the return value covers a narrower
range than the data itself, then the youngest age
group in the return value includes ASFRs from younger age
groups in the data, and the oldest age group in the return
value includes ASFRs from older age groups in the data.
For instance, if the classification in the return value
starts at age 15, but the data starts at age 12,
the ASFR for age 15 includes ASFRs from ages 12, 13, and 14.

This shifting of ASFRs is common in analyses of fertility.
}

\examples{
data <- data_ssvd_hfd(asfr_subset)
data
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Prepare data on age-specific mortality
from the Human Mortality Database
\item \code{\link[=data_ssvd_lfp]{data_ssvd_lfp()}} Prepare data on labour force participation
from the OCED
}
}
