% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himd.R
\name{data_ssvd_himd}
\alias{data_ssvd_himd}
\title{Prepare Data from Human Internal Migration Database}
\usage{
data_ssvd_himd(
  zipfile,
  time_interval = 1,
  measure_type = c("rate", "prob"),
  n_comp = 5
)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Internal Migration Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{time_interval}{Length of interval over
which values are calculated. Choices are
\code{1} or \code{5}. Default is \code{1}.}

\item{measure_type}{\code{"prob"} or \code{"rate"}.
Whether values are to be
treated as probabilities or
as rates. See \code{\link[=data_ssvd_himd]{data_ssvd_himd()}} for details.
Default is \code{"rate"}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Process data on age-specific migration probabilities
from the Human Internal Migration Database.
}
\section{Usage}{
\strong{Step 1: Download data}

Download the data from
Dyrting, S. (2024, October 23). \emph{Data from: Estimating
Complete Migration Probabilities from Grouped Data}
at https://osf.io/vmrfk/.
The data is in the "Migration"
folder under the "Files" tab.
The facility for downloading all the files at once
does not appear to be working, but CSV files can
be downloaded one by one. Zip the folder
containing these CSV files.

**Step 2: Call function \code{data_ssvd_himd()}

Create three datasets:

\if{html}{\out{<div class="sourceCode">}}\preformatted{himd_data_rate <- data_svd_himd(zipfile = "himd.zip",
                                time_interval = 1,
                                measure_type = "rate")
himd_data_prob1 <- data_svd_himd(zipfile = "himd.zip",
                                 time_interval = 1,
                                 measure_type = "prob")
himd_data_prob5 <- data_svd_himd(zipfile = "himd.zip",
                                 time_interval = 5,
                                 measure_type = "prob")
}\if{html}{\out{</div>}}
}

\section{Rates or probabilities}{
The original data are probabilities.
However, \code{data_ssvd_himd()} (and \code{\link[=coef_himd]{coef_himd()}})
allow values where \code{time_interval} is
1 to be to treated as rates. If more than
one change in residence in a year is unusual,
then one-year probabilities and should
in fact be very similar.

When values are being treated as probabilities,
a logit transform is applied before
carrying out the SVD. When they are being
treated as rates, a log transform is applied.
}

\examples{
zipfile <- system.file("extdata", "himd_subset.zip",
                       package = "bssvd")
data <- data_ssvd_himd(zipfile)
data
}
\seealso{
\itemize{
\item \code{\link[=coef_hfd]{coef_hfd()}} Obtain time series
of coefficients
for Human Internal Migration Database data
\item \code{\link[=tidy_hmd]{tidy_hmd()}} Format data from the
Human Internal Migration Database into
a tidy data frame
}
}
