% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{data_ssvd_hmd}
\alias{data_ssvd_hmd}
\title{Prepare Data from Human Mortality Database}
\usage{
data_ssvd_hmd(
  zipfile,
  n_comp = 5,
  date = c("2025-09-25", "2024-02-26"),
  year_min = NULL,
  eps = 1e-05
)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}

\item{date}{Date on which data produced.
Choices are \code{"2025-09-25"}, \code{"2024-02-26"}.
(Class is character, not Date.)}

\item{year_min}{Only include data for
\code{year_min} onwards. Ignored if
\code{year_min} is \code{NULL} (the default).}

\item{eps}{Floor for rates.
Default is \code{0.00001}.}
}
\value{
A tibble
}
\description{
Process data on age-specific mortality rates
from a zipped file downloaded from the
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}
so that it is ready to created a scaled SVD.
}
\section{Usage}{

\strong{Step 1: Download data}

Register or log in at the Human Mortality Database, and go to page
\href{https://www.mortality.org/Data/ZippedDataFiles}{Downloading the HMD in zipped data files}.
Go to the "Previous Versions" table at the bottom of the page, and
download a file from the "Statistics" column, eg file

https://www.mortality.org/File/Download/hmd.v6/zip/all_hmd/hmd_statistics_20240226.zip

\strong{Step 2: Call function \code{data_ssvd_hmd()}}

Put the file into the working directory, and supply the name of the file
to function \code{data_ssvd_hmd()}, eg

\if{html}{\out{<div class="sourceCode">}}\preformatted{hmd_data <- data_ssvd_hmd("hmd_statistcs_20240226")
}\if{html}{\out{</div>}}
}

\section{Truncation and transformation}{


Rates below \code{eps} are shifted to \code{eps},
and all rates are log-transformed,
before the SVD is applied.
}

\examples{
zipfile <- system.file("extdata", "hmd_statistics_20250925_subset.zip",
                       package = "bssvd")
data <- data_ssvd_hmd(zipfile)
data
}
\seealso{
\itemize{
\item \code{\link[=coef_hfd]{coef_hfd()}} Obtain time series
of SVD coefficients
for Human Mortality Database data
\item \code{\link[=tidy_hmd]{tidy_hmd()}} Format data from the
Human Mortality Database into
a tidy data frame
}
}
