% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfp.R
\name{data_ssvd_lfp}
\alias{data_ssvd_lfp}
\title{Prepare OECD Data on Labour Force Participation Rates}
\usage{
data_ssvd_lfp(data, age_max = 75, n_comp = 5, year_min = NULL)
}
\arguments{
\item{data}{A data frame containing OECD data.}

\item{age_max}{The upper limit of the
oldest closed age group. Default is \code{75}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}

\item{year_min}{Only include data for
\code{year_min} onwards. Ignored if
\code{year_min} is \code{NULL} (the default).}
}
\value{
A tibble
}
\description{
Process labour force participation data downloaded from
the \href{https://data-explorer.oecd.org}{OECD Data Explorer}
so it is ready to create a scaled SVD.
}
\section{Usage}{

\strong{Step 1: Download data}

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(rsdmx)
url <- paste("https://sdmx.oecd.org/public/rest/data",
             "OECD.ELS.SAE,DSD_LFS@DF_LFS_INDIC,1.1",
             "all?dimensionAtObservation=AllDimensions",
             sep = "/")
lfp_sdmx <- rsdmx::readSDMX(url)    ## can be slow
lfp_df <- as.data.frame(lfp_sdmx)   ## can also be slow
}\if{html}{\out{</div>}}

\strong{Step 2: Call function \code{data_ssvd_lfp()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{lfp_data <- data_ssvd_lfp(data)
}\if{html}{\out{</div>}}
}

\examples{
data <- data_ssvd_lfp(oecd_lfp_subset)
data
}
\seealso{
\itemize{
\item \code{\link[=tidy_lfp]{tidy_lfp()}} Format OECD labour force
participation data into a tidy data frame.
\item \code{\link[=coef_lfp]{coef_lfp()}} Obtain time series
of coefficients for OECD labour force
participation data
}
}
