% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmd.R
\name{data_ssvd_wmd}
\alias{data_ssvd_wmd}
\title{Process World Marriage Data}
\usage{
data_ssvd_wmd(file, n_comp = 5, status = c("current", "ever"), eps = 1e-05)
}
\arguments{
\item{file}{Path to excel spreadsheet
holding World Marriage Data}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}

\item{status}{\code{"current"} (the default)
or \code{"ever"}. Whether to report results for
the marriage staus "currently married"
or "ever-married". Note that both statuses
include consensual and de facto unions,
in addition to legal marriages.}

\item{eps}{Parameter for truncating
probabililities or rates.
Default is \code{0.00001}.}
}
\value{
A tibble
}
\description{
Process marriage data downloaded from
the UN Population Division's
\href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Marriage Data},
so it is ready to create a scaled SVD.
}
\section{Usage}{

\strong{Step 1: Download data}

Go to to page \href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Marriage Data}.
Go to the data tab on the right. Download the most
recent excel spreadsheet (which as of October 2025 is called
\code{undesa_pd_2019_wmd_marital_status.xlsx}).

\strong{Step 2: Call function \code{data_ssvd_wmd()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{wmd_data <- data_ssvd_wmd("undesa_pd_2019_wmd_marital_status.xlsx")
}\if{html}{\out{</div>}}
}

\section{coef_wmd Truncation and transformation}{

}

\examples{
file <- system.file(
  "extdata",
  "undesa_pd_2019_wmd_marital_status_subset.xlsx",
  package = "bssvd"
)
coef_wmd(file)
}
\seealso{
\itemize{
\item \code{\link[=coef_wmd]{coef_wmd()}} Obtain time series
of SVD coefficients
from World Marriage Data
\item \code{\link[=tidy_wmd]{tidy_wmd()}} Format World Marriage Data
into a tidy data frame.
}
}
