% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd.R
\name{hfd_tidy}
\alias{hfd_tidy}
\title{Tidy HFD Age-Specific Fertility Data}
\usage{
hfd_tidy(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A tibble.
}
\description{
Tidy data frame containing data from
the ASFR row of the "By statistic" table of
\href{https://www.humanfertility.org/Data/ZippedDataFiles}{ZippedDataFiles}.
}
\examples{
hfd_tidy(asfr)
}
\seealso{
\itemize{
\item \code{\link[=ssvd_hfd]{ssvd_hfd()}} Tidy data, and convert it into
and an SVD object.
\item \code{\link[=lfp_tidy]{lfp_tidy()}} Tidy OECD data on labour force participation.
}
}
