% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfp.R
\name{lfp_tidy}
\alias{lfp_tidy}
\title{Tidy OECD Labor Force Participation Data}
\usage{
lfp_tidy(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A tibble.
}
\description{
Tidy data frame containing data downloaded from
the Labour Force Indicators table of the
\href{https://data-explorer.oecd.org}{OECD Data Explorer}.
}
\examples{
data_tidy <- lfp_tidy(oecd_lfp)
}
\seealso{
\itemize{
\item \code{\link[=ssvd_lfp]{ssvd_lfp()}} Tidy data, and convert it into
and an SVD object.
}
}
