% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{marital_status}
\alias{marital_status}
\title{Marital Status}
\format{
A tibble with 35 rows and with
columns \code{marital_status} and \code{currently_married},
and \code{ever_married}.
}
\source{
Marital statuses derived from data in
file \code{undesa_pd_2019_wmd_marital_status.xlsx}
downloaded from the United Nations Population Division
website https://www.un.org/development/desa/pd/data/world-marriage-data
on 10 October 2025.
}
\usage{
marital_status
}
\description{
Mapping between reported marital statuses
reported in the World Marriage Data database
and variables 'currently_married' and 'ever_married'.
}
\keyword{datasets}
