% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oecd_lfp}
\alias{oecd_lfp}
\title{OECD data on Labour Force Participation}
\format{
A tibble with 12069 rows and the following
columns `TIME_PERIOD`, `REF_AREA`,
`MEASURE`,`UNIT_MEASURE`, `SEX`, `AGE`,
`LABOUR FORCE STATUS`, `obsValue`, `DECIMALS`,
`OBS_STATUS`.
}
\source{
Downloaded from table
"Employment and unemployment by five-year
age group and sex - indicators" from
[OECD Data Explorer](https://data-explorer.oecd.org).
Accessed 21 April 2024.
Only a subset of the full dataset is included
in `oecd_lpr`.
}
\usage{
oecd_lfp
}
\description{
Data on labour force participation rates
by age and sex, for
multiple countries, starting in 1960,
from the OECD Data Explorer database.
}
\details{
The labour force participation rate is the
number of people in the labour force (ie the
number of people who are currently employed
or unemployed) as a proportion of the population.
}
\keyword{datasets}
