% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oecd_lfp_subset}
\alias{oecd_lfp_subset}
\title{OECD Data on Labour Force Participation}
\format{
A tibble with 12,069 rows and with
columns \code{TIME_PERIOD}, \code{REF_AREA},
\code{MEASURE},\code{UNIT_MEASURE}, \code{SEX}, \code{AGE},
\verb{LABOUR FORCE STATUS}, \code{obsValue}, \code{DECIMALS},
\code{OBS_STATUS}.
}
\source{
Downloaded from table
"Employment and unemployment by five-year
age group and sex - indicators" from
\href{https://data-explorer.oecd.org}{OECD Data Explorer}.
Accessed 21 April 2024.
Only a subset of the data is included
in \code{oecd_lfp_subset}.
}
\usage{
oecd_lfp_subset
}
\description{
Data on labour force participation rates
by age and sex, for
multiple countries in 2010--2012,
extracted from the OECD Data Explorer database.
}
\details{
The labour force participation rate is the
number of people in the labour force (ie the
number of people who are currently employed
or unemployed) as a proportion of the population.
}
\keyword{datasets}
