% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{percent_variance}
\alias{percent_variance}
\title{Calculate the Percent of Variance Explained
by Components in SVD}
\usage{
percent_variance(
  data,
  measure,
  cols = c("country", "time"),
  transform = c("log", "logit"),
  n_comp = 5,
  eps = 1e-05
)
}
\arguments{
\item{data}{A data frame}

\item{measure}{Name of the rate or proportion
being modelled.}

\item{cols}{Names of the categorical variables
used to distinguish different age profiles.}

\item{transform}{Transformation to apply
to the data: \code{"log"} (the default) or \code{"logit"}.}

\item{n_comp}{Number of components to
give percentages for. Default is \code{5}.}

\item{eps}{Parameter controlling truncation.
See below for details. Default is \code{0.00001}.}
}
\value{
A numeric vector of length \code{n_comp}.
}
\description{
Calculate what percent of variance in a dataset
can be explained by the first \code{n_comp} components
of a singular value decompositon. The dataset should
include age, a single sex or gender category,
and one or more classifying variables.
}
\details{
The dataset is transformed
before the SVD is applied. If the transformation
is \code{"log"}, then zeros are moved upwards to the
lowest non-zero value appearing in the data
before the transformation is applied.
If the transformation is \code{"logit"}, then
zeros are moved upwards and ones are
moved downward.

\code{data} must contain a variable called \code{"age"}.
It must also contain the variables named by
the \code{measure} and \code{cols} arguments. The age
and \code{cols} arguments must uniquely identify
all cells in the dataset.

Before the log or logit transforms are applied
all values less than \code{eps} are changed to \code{eps}.
Before the logit transform is applied, all values
greater than \code{1 - eps} are changed to \code{1 - eps}.
}
