% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{tidy_csa}
\alias{tidy_csa}
\title{Tidy World Marriage Data}
\usage{
tidy_csa(data)
}
\value{
A tibble
}
\description{
Tidy data downloaded from
the UN Population Division's
\href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Migration Data},
database.
}
\examples{
tidy_wmd(un_csa_subset)
}
\seealso{
\itemize{
\item \code{\link[=coef_wmd]{coef_wmd()}} Obtain time series
of SVD coefficients
from World Marriage Data
\item \code{\link[=data_ssvd_wmd]{data_ssvd_wmd()}} Put World Marriage Data
into the format required by
the \code{ssvd()} function in \pkg{bage}
}
}
