% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd.R
\name{tidy_hfd}
\alias{tidy_hfd}
\title{Tidy HFD Age-Specific Fertility Data}
\usage{
tidy_hfd(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A tibble.
}
\description{
Tidy a data frame containing data from
the ASFR row of the "By statistic" table of
\href{https://www.humanfertility.org/Data/ZippedDataFiles}{ZippedDataFiles}.
}
\examples{
tidy_hfd(asfr_subset)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hfd]{data_ssvd_hfd()}}Put data
from the Human Fertility Database
into the format required by the
\code{ssvd()} function in \code{bage}.
\item \code{\link[=coef_hfd]{coef_hfd()}} Obtain time series
of coefficients
from Human Fertility Database data
}
}
