% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himd.R
\name{tidy_himd}
\alias{tidy_himd}
\title{Tidy HIMD Age-Specific Internal Migration Data}
\usage{
tidy_himd(zipfile, time_interval = 1)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Internal Migration Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{time_interval}{Length of interval over
which values are calculated. Choices are
\code{1} or \code{5}. Default is \code{1}.}
}
\value{
A tibble
}
\description{
Put data on age-specific migration probabilities
from a zipped file downloaded from the
\href{https://osf.io/vmrfk/}{Human Internal Migration Database}
into a tidy data frame.
}
\examples{
zipfile <- system.file("extdata", "himd_20241023_subset.zip",
                       package = "bssvd")
tidy_himd(zipfile)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_himd]{data_ssvd_himd()}} Put data
from the Human Internal Migration Database
into the format required by
the \code{ssvd()} function in \pkg{bage}
\item \code{\link[=coef_hmd]{coef_hmd()}} Obtain time series
of SVD coefficients
for Human Internal Migration Database data
}
}
