% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{tidy_hmd}
\alias{tidy_hmd}
\title{Tidy HMD Age-Specific Mortality Data}
\usage{
tidy_hmd(zipfile, date = c("2025-09-25", "2024-02-26"), year_min = NULL)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{year_min}{Only include data for
\code{year_min} onwards. Ignored if
\code{year_min} is \code{NULL} (the default).}
}
\value{
A tibble
}
\description{
Put data on age-specific mortality rates
from a zipped file downloaded from the
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}
into a tidy data frame.
}
\examples{
zipfile <- system.file("extdata", "hmd_statistics_subset.zip",
                       package = "bssvd")
tidy_hmd(zipfile)
tidy_hmd(zipfile,
         date = "2024-02-26",
         year_min = 1950)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Put data
from the Human Mortality Database
into the format required by
the \code{ssvd()} function in \code{bage}
\item \code{\link[=coef_hmd]{coef_hmd()}} Obtain time series
of coefficients
for Human Mortality Database data
}
}
