% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfp.R
\name{tidy_lfp}
\alias{tidy_lfp}
\title{Tidy OCED Labour Force Participation Data}
\usage{
tidy_lfp(data, year_min = NULL)
}
\arguments{
\item{data}{A data frame containing OECD data.}

\item{year_min}{Only include data for
\code{year_min} onwards. Ignored if
\code{year_min} is \code{NULL} (the default).}
}
\value{
A tibble
}
\description{
Put data on labour force participation
data downloaded from
the \href{https://data-explorer.oecd.org}{OECD Data Explorer}
into a tidy data frame.
}
\examples{
tidy_lfp(oecd_lfp_subset)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_lfp]{data_ssvd_lfp()}} Put OECD labour
force participation data
into the format required by
the \code{ssvd()} function in \code{bage}
\item \code{\link[=tidy_lfp]{tidy_lfp()}} Format OECD labour force
participation data into a tidy data frame.
\item \code{\link[=coef_lfp]{coef_lfp()}} Obtain time series
of coefficients for OECD labour force
participation data
}
}
