% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmd.R
\name{tidy_wmd}
\alias{tidy_wmd}
\title{Tidy World Marriage Data}
\usage{
tidy_wmd(file, status = c("current", "ever"))
}
\arguments{
\item{file}{Path to excel spreadsheet
holding World Marriage Data}

\item{status}{\code{"current"} (the default)
or \code{"ever"}. Whether to report results for
the marriage staus "currently married"
or "ever-married". Note that both statuses
include consensual and de facto unions,
in addition to legal marriages.}
}
\value{
A tibble
}
\description{
Tidy data downloaded from
the UN Population Division's
\href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Marriage Data},
database.
}
\examples{
file <- system.file(
  "extdata",
  "undesa_pd_2019_wmd_marital_status_subset.xlsx",
  package = "bssvd"
)
tidy_wmd(file)
}
\seealso{
\itemize{
\item \code{\link[=coef_wmd]{coef_wmd()}} Obtain time series
of SVD coefficients
from World Marriage Data
\item \code{\link[=data_ssvd_wmd]{data_ssvd_wmd()}} Put World Marriage Data
into the format required by
the \code{ssvd()} function in \pkg{bage}
}
}
