% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{un_csa_subset}
\alias{un_csa_subset}
\title{Census Data on School Attendance}
\format{
A tibble with 7668 rows and with columns
"Country or Area", "Year", "Area", "Sex",
"Age", "School attendance", "Record Type",
"Reliability", "Source Year", "Value",
and "Value Footnotes"
}
\source{
Downloaded from table
"Population 5 to 24 years of age by school
attendance, sex and urban/rural residence"
from the "Population Censuses' Datasets (1995 - Present)"
in the \href{https://unstats.un.org/unsd/demographic-social/products/dyb/index.cshtml}{Demographic Statistics Database}. Accessed 22 October 2025.
}
\usage{
un_csa_subset
}
\description{
Data on school attendance rates,
by age and sex, from censuses,
assembled by the United Nations
Statistics Division.
}
\keyword{datasets}
