% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssvd-package.R
\docType{package}
\name{bssvd-package}
\alias{bssvd}
\alias{bssvd-package}
\title{bssvd: Make Scaled SVD Objects}
\description{
Based on data from international demographic databases, build objects of class "bage_ssvd", to be used with package 'bage'. 'bssvd' is a developer-focused package, and would not normally be used directly for data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayesiandemography.github.io/bssvd/}
}

}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

}
\keyword{internal}
