% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd.R
\name{coef_hfd}
\alias{coef_hfd}
\title{Obtain Coefficients from Scaled SVD of HFD Data}
\usage{
coef_hfd(data, n_comp = 5)
}
\arguments{
\item{data}{A data frame containing HFD data.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Obtain time series of coefficients from a
scaled SVD of
\href{https://www.humanfertility.org/Home/Index}{Human Fertility Database}
data.
}
\details{
Obtain a scaled version of the \eqn{U} matrix,
produced as part of the SVD of log age-specific
fertility rates from the HFD.
}
\examples{
coef_hfd(asfr_subset)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hfd]{data_ssvd_hfd()}} Prepare data on fertility
from the Human Fertility Database
\item \code{\link[=coef_hmd]{coef_hmd()}} Obtain coefficients
for Human Mortality Database data
\item \code{\link[=coef_lfp]{coef_lfp()}} Obtain coefficients
for OECD Labor Force Participation data
}
}
