% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{coef_hmd}
\alias{coef_hmd}
\title{Obtain Coefficients from Scaled SVD of HMD Data}
\usage{
coef_hmd(
  zipfile,
  n_comp = 5,
  date = c("2025-09-25", "2024-02-26"),
  year_min = NULL
)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}

\item{year_min}{Only include data for
\code{year_min} onwards. Ignored if
\code{year_min} is \code{NULL} (the default).}

\item{Date}{on which data produced.
Choices are \code{"2025-09-25"}, \code{"2024-02-26"}.
(Class is character, not Date.)}
}
\value{
A tibble
}
\description{
Obtain time series of coefficients from a
scaled SVD of data from the
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}.
The coefficients are a
scaled version of the \eqn{U} matrix
from the SVD.
}
\examples{
zipfile <- system.file("extdata", "hmd_statistics_subset.zip",
                       package = "bssvd")
coef_hmd(zipfile)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Put data
from the Human Mortality Database
into the format required by
the \code{ssvd()} function in \code{bage}
\item \code{\link[=tidy_hmd]{tidy_hmd()}} Format data from the
Human Mortality Database into
a tidy data frame
}
}
