% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{coef_hmd}
\alias{coef_hmd}
\title{Obtain Coefficients from Scaled SVD of HMD Data}
\usage{
coef_hmd(zipfile, n_comp = 5)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Obtain time series of coefficients from a
scaled SVD of
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}
data.
}
\details{
Obtain a scaled version of the \eqn{U} matrix,
produced as part of the SVD of log
mortality rates from the HMD.
The mortality rates are for single years of age
up to age 110+.
}
\examples{
zipfile <- system.file("extdata", "hmd_statistics_subset.zip",
                       package = "bssvd")
coef_hmd(zipfile)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Prepare data on fertility
from the Human Mortality Database
\item \code{\link[=coef_hfd]{coef_hfd()}} Obtain coefficients
for Human Fertility Database data
\item \code{\link[=coef_lfp]{coef_lfp()}} Obtain coefficients
for OECD Labor Force Participation data
}
}
