% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfp.R
\name{coef_lfp}
\alias{coef_lfp}
\title{Obtain Coefficients from Scaled SVD of OCED
Labour Force Participation Data}
\usage{
coef_lfp(data, n_comp = 5)
}
\arguments{
\item{data}{A data frame containing OECD data.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble
}
\description{
Obtain coefficients from a
scaled SVD of OCED labour force participation
data downloaded from
the \href{https://data-explorer.oecd.org}{OECD Data Explorer}.
The coefficients are a
scaled version of the \eqn{U} matrix
from the SVD.
}
\examples{
coef_lfp(oecd_lfp_subset)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Prepare data on fertility
from the Human Mortality Database
\item \code{\link[=coef_hfd]{coef_hfd()}} Obtain coefficients
for Human Fertility Database data
\item \code{\link[=coef_hmd]{coef_hmd()}} Obtain coefficients
for Human Mortality Database data
}
}
