% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{data_ssvd_csa}
\alias{data_ssvd_csa}
\title{Prepare Census School Attendance Data}
\usage{
data_ssvd_csa(data, n_comp = 5, eps = 1e-05)
}
\arguments{
\item{data}{Data frame with attendance data.
(Data from multiple files combined, but otherwise
unprocessed).}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}

\item{eps}{Parameter for truncating
probabililities.
Default is \code{0.00001}.}
}
\value{
A tibble
}
\description{
Process school attendance data,
so it is ready to create a scaled SVD.
}
\section{Usage}{

\strong{Step 1: Download data}

Go to to the
\href{https://unstats.un.org/unsd/demographic-social/products/dyb/index.cshtml#censusdatasets}{Population Censuses' Datasets} database, then to the table
"Population 15 years of age and over
by educational attainment, age and sex".
Go to the download tab, and download. Only 100,000 rows
can be downloaded at one time
so this needs to be done in multiple steps.

\strong{Step 2: Combine datasets}

Read in the downloadeded files, and combine into a single
data frame.

\strong{Step 3: Call function \code{data_ssvd_csa()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{csa_data <- data_ssvd_ssvd(data)
}\if{html}{\out{</div>}}
}

\section{coef_csa Truncation and transformation}{

}

\examples{
data_ssvd_csa(un_csa_subset)
}
\seealso{
\itemize{
\item \code{\link[=coef_csa]{coef_csa()}} Obtain time series
of SVD coefficients
from census cchool attendance data
\item \code{\link[=tidy_csa]{tidy_csa()}} Format census school attendance
data into a tidy data frame.
}
}
