% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{data_ssvd_hmd}
\alias{data_ssvd_hmd}
\title{Prepare Data from Human Mortality Database}
\usage{
data_ssvd_hmd(zipfile, n_comp = 5)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble with the format required by
\code{bage::ssvd()}.
}
\description{
Process data on age-specific mortality rates
from a zipped file downloaded from the
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}.
}
\section{Usage}{

\strong{Step 1: Download data}

Register or log in at the Human Mortality Database, and go to page
\href{https://www.mortality.org/Data/ZippedDataFiles}{Downloading the HMD in zipped data files}.
Go to the "Previous Versions" table at the bottom of the page, and
download a file from the "Statistics" column, eg file

https://www.mortality.org/File/Download/hmd.v6/zip/all_hmd/hmd_statistics_20240226.zip

\strong{Step 2: Call function 'data_ssvd_hmd'}

Put the file into the working directory, and supply the name of the file
to function \code{data_ssvd_hmd()}, eg

\if{html}{\out{<div class="sourceCode">}}\preformatted{hmd_data <- data_ssvd_hmd("hmd_statistcs_20240226")
}\if{html}{\out{</div>}}
}

\examples{
zipfile <- system.file("extdata", "hmd_statistics_subset.zip",
                       package = "bssvd")
data <- data_ssvd_hmd(zipfile)
data
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hfd]{data_ssvd_hfd()}} Prepare data on fertility
from the Human Fertility Database
\item \code{\link[=data_ssvd_lfp]{data_ssvd_lfp()}} Prepare data on labour force participation
from the OCED
}
}
