% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{data_ssvd_hmd}
\alias{data_ssvd_hmd}
\title{Prepare Data frome Human Mortality Database}
\usage{
data_ssvd_hmd(zipfile, n_comp = 10)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
A path name that is handled by \code{\link[utils:unzip]{utils::unzip()}}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{10}.}
}
\value{
A tibble with the format required by
function \code{bage::ssvd()}.
}
\description{
Process data on age-specific mortality rates
from a zipped file downloaded from the
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}.
}
\section{Usage}{

\strong{Step 1: Download data}

Register or log in at the Human Mortality Database, and go to page
\href{https://www.mortality.org/Data/ZippedDataFiles}{Downloading the HMD in zipped data files}.
Go to the "Previous Versions" table at the bottom of the page, and
download a file from the "Statistics" column, eg file

https://www.mortality.org/File/Download/hmd.v6/zip/all_hmd/hmd_statistics_20240226.zip

\strong{Step 2: Call function 'svd_hmd'}

Put the file into the working directory, and supply the name of the file
to function \code{data_ssvd_hmd()}, eg

\if{html}{\out{<div class="sourceCode">}}\preformatted{data_hmd <- data_ssvd_hmd("hmd_statistcs_20240226")
}\if{html}{\out{</div>}}
}

\seealso{
\itemize{
\item \code{\link[=data_ssvd_hfd]{data_ssvd_hfd()}} Prepare data on age-specific fertility rates
from Human Fertility Database
\item \code{\link[=data_ssvd_lfs]{data_ssvd_lfs()}} Prepare data on labour force participation
rates from the OCED
\item \code{\link[bage:ssvd]{bage::ssvd()}} Create Scaled SVD object
}
}
