% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfp.R
\name{data_ssvd_lfp}
\alias{data_ssvd_lfp}
\title{Prepare OECD Data on Labour Force Participation Rates}
\usage{
data_ssvd_lfp(data, age_max = 75, n_comp = 5)
}
\arguments{
\item{data}{A data frame containing OECD data.}

\item{age_max}{The upper limit of the
oldest closed age group. Default is \code{75}.}

\item{n_comp}{Number of SVD components
to include in result. Default is \code{5}.}
}
\value{
A tibble with the format required by
function \code{bage::ssvd()}.
}
\description{
Process labour force participation data downloaded from
the \href{https://data-explorer.oecd.org}{OECD Data Explorer}.
}
\section{Usage}{

\strong{Step 1: Download data}

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(rsdmx)
url <- paste("https://sdmx.oecd.org/public/rest/data",
             "OECD.ELS.SAE,DSD_LFS@DF_LFS_INDIC,1.1",
             "all?dimensionAtObservation=AllDimensions",
             sep = "/")
lfp_sdmx <- rsdmx::readSDMX(url)    ## can be slow
data <- as.data.frame(lfp_sdmx)     ## can also be slow
}\if{html}{\out{</div>}}

\strong{Step 2: Call function 'data_ssvd_lfp()'}

\if{html}{\out{<div class="sourceCode">}}\preformatted{lfp_data <- data_ssvd_lfp(data)
}\if{html}{\out{</div>}}
}

\examples{
lfp_data_small <- data_ssvd_lfp(oecd_lfp)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hfd]{data_ssvd_hfd()}} Prepare data on age-specific fertility rates
from Human Fertility Database.
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Prepare data on age-specific mortality rates
from Human Mortality Database.
\item \code{\link[bage:ssvd]{bage::ssvd()}} Create Scaled SVD object
}
}
