% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd.R
\name{hfd_tidy}
\alias{hfd_tidy}
\title{Tidy HFD Age-Specific Fertility Data}
\usage{
hfd_tidy(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A tibble.
}
\description{
Tidy a data frame containing data from
the ASFR row of the "By statistic" table of
\href{https://www.humanfertility.org/Data/ZippedDataFiles}{ZippedDataFiles}.
}
\examples{
hfd_tidy(asfr_subset)
}
\seealso{
\itemize{
\item \code{\link[=data_ssvd_hmd]{data_ssvd_hmd()}} Prepare data on mortality
from Human Mortality Database
\item \code{\link[=data_ssvd_lfp]{data_ssvd_lfp()}} Prepare data on labour force participation
from the OCED
}
}
