% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_variance.R
\name{percent_variance}
\alias{percent_variance}
\title{Calculate the Percent of Variance Explained
by Components in SVD}
\usage{
percent_variance(
  data,
  transform,
  measure = "value",
  cols = c("country", "time"),
  n_comp = 5,
  eps = 1e-05
)
}
\arguments{
\item{data}{A data frame}

\item{transform}{Transformation to apply
to the data: \code{"log"}, \code{"logit"}, or \code{"none"}.}

\item{measure}{Name of the rate or proportion
being modelled. Default is \code{"value"}.}

\item{cols}{Names of the categorical variables
used to distinguish different age profiles.}

\item{n_comp}{Number of components to
give percentages for. Default is \code{5}.}

\item{eps}{Parameter controlling truncation.
See below for details. Default is \code{0.00001}.}
}
\value{
A numeric vector of length \code{n_comp}.
}
\description{
Calculate the percent of variance in a dataset
that can be explained by each component
of a singular value decomposition. The dataset should
include age, a single sex or gender category,
and one or more classifying variables.
}
\details{
\code{data} must contain a variable called \code{"age"}.
It must also contain the variables named by
the \code{measure} and \code{cols} arguments. The age
and \code{cols} arguments must uniquely identify
all cells in the dataset.

When \code{transform} is \code{"log"} or \code{"logit"},
\code{percent_variance()} converts any
values for the measure variable
that are less than \code{eps} to
values greater than or equal to \code{eps} before
transforming. The converted value is based on
on a main effects model.

When \code{transform} is \code{"logit"},
\verb{precent_variange() converts any values greater than }1-eps\verb{to values less than or equal to}1-eps`
before transforming. The converted value
is based on main effects model.
}
\examples{
asfr_subset |>
  tidy_hfd() |>
  percent_variance(transform = "log")
}
