% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfp.R
\name{ssvd_lfp}
\alias{ssvd_lfp}
\title{Create a Scaled SVD Object from OECD Labour Force Participation Data}
\usage{
ssvd_lfp(data, age_max = 75, n_comp = 5)
}
\arguments{
\item{data}{A data frame containing OECD data.}

\item{age_max}{The upper limit of the
oldest closed age group. Default is `75`.}

\item{n_comp}{Number of SVD components
to include in result. Default is `5`.}
}
\value{
An object of class [`"bage_ssvd"`][ssvd()]
}
\description{
Create an object of class [`"bage_ssvd"`][bage::ssvd()]
from labour force participation data downloaded from
the [OECD Data Explorer](https://data-explorer.oecd.org).
}
\section{Usage}{

**Step 1: Download data**

```
library(rsdmx)
url <- paste("https://sdmx.oecd.org/public/rest/data",
             "OECD.ELS.SAE,DSD_LFS@DF_LFS_INDIC,1.1",
             "all?dimensionAtObservation=AllDimensions",
             sep = "/")
lfp_sdmx <- rsdmx::readSDMX(url)    ## can be slow
lfp_df <- as.data.frame(lfp_sdmx)   ## can also be slow
```

**Step 2: Call function 'svd_lfp'**

```
LFP <- svd_lfp(data)
```
}

\seealso{
- [bage::SVD()], [bage::SVDS()], [bage::ESVD()], [bage::ESVDS()]
  to use an object of class
  [`"bage_ssvd"`][bage::ssvd()] in a prior.
}
